requirements.txt
cycler==0.10.0
dlib==19.19.0
imutils==0.5.3
kiwisolver==1.1.0
matplotlib==3.1.3
numpy==1.18.1
opencv-contrib-python==4.2.0.32
Pillow==7.0.0
pyparsing==2.4.6
python-dateutil==2.8.1
scipy==1.4.1
six==1.14.0
